initial refactor for moving GoToAprilTag to Movement class so we can use the same code to auto align to the backdrop via button in teleop
